Move single-use util functions into routes
