+++ app/routes/web/direct.py
@@ -68,6 +68,42 @@ def online_beatmap(set: DBBeatmapset, post_id: int = 0) -> str:
         str(post_id or 0),
     ])
 
+def update_osz_filesize(set_id: int, has_video: bool = False) -> None:
+    updates = {}
+
+    if has_video:
+        updates['osz_filesize_novideo'] = get_osz_size(
+            set_id,
+            no_video=True
+        )
+
+    updates['osz_filesize'] = get_osz_size(
+        set_id,
+        no_video=False
+    )
+
+    beatmapsets.update(set_id, updates)
+
+def get_osz_size(set_id: int, no_video: bool = False) -> int:
+    r = app.session.requests.head(
+        f'https://osu.direct/api/d/{set_id}'
+        f'{"noVideo=" if no_video else ""}'
+    )
+
+    if not r.ok:
+        app.session.logger.error(
+            f"Failed to get osz size: {r.status_code}"
+        )
+        return 0
+
+    if not (filesize := r.headers.get('content-length')):
+        app.session.logger.error(
+            "Failed to get osz size: content-length header missing"
+        )
+        return 0
+
+    return int(filesize)
+
 @router.get('/osu-search.php')
 def search(
     session: Session = Depends(app.session.database.yield_session),
@@ -202,7 +238,7 @@ def pickup_info(
     )
 
     if not beatmapset.osz_filesize:
-        utils.update_osz_filesize(
+        update_osz_filesize(
             beatmapset.id,
             beatmapset.has_video
         )
